Move AttackRequirement
